\name{compare_vectors}
\alias{compare_vectors}
\title{
  Compare Vectors
}
\description{
  Compare numeric or logical vectors.
}
\usage{
compare_vectors(..., sep = "")
}
\arguments{
  \item{\dots}{
    vectors of the same length
  }
  \item{sep}{
    a string
  }
}
\details{

  Very experimental.

}
\value{

  typically a list

}
%% \references{
%% %% ~put references to the literature/web site here ~
%% }
\author{
  Enrico Schumann
}
\seealso{
  \code{\link{random_vector}}
}
\examples{
## TODO
1+1
}
