\name{neighbours-package}
\alias{neighbours-package}
\alias{neighbours}
\docType{package}
\title{
\packageTitle{neighbours}
}
\description{
\packageDescription{neighbours}
}
\details{

The package currently provides only one function: \code{\link{neighbourfun}}.

}
\author{
\packageAuthor{neighbours}

Maintainer: \packageMaintainer{neighbours}
}
\references{

  Gilli, M., Maringer, D. and Schumann, E. (2011) \emph{Numerical
  Methods and Optimization in Finance}. Elsevier.
  \url{https://www.elsevier.com/books/numerical-methods-and-optimization-in-finance/gilli/978-0-12-375662-6}

  Schumann, E. (2018) Financial Optimisation with R (\pkg{NMOF} Manual).
  \url{http://enricoschumann.net/NMOF.htm#NMOFmanual}

}
\keyword{ package }
